API

Theseus.combineSorted — Method

Returns a sorted matrix of points based on their indices in the original matrix. This allows for changes to some elements of the matrix in a differentiable way since zygote does not support mutating operations and the typical assignment operator mutates.

Theseus.crossEntropyLoss — Method

Weighted cross entropy loss, best used with a plain mask vector of weights.

Theseus.crossEntropyLoss — Method

Cross entropy loss function.

Theseus.forceVar — Method

Reduce the difference between the maximum and minimum forces in the network. From Schek theorem 2.

Theseus.lenVar — Method

Compute difference between the maximum and minimum lengths of the edges in the network.

Theseus.logSumExp — Method

LogSumExp is the multivariable generalization of the logistic function.

Theseus.logisticFunc — Method

The derivative of the softplus function is the logistic function. A scaling parameter k can be introduced to make the inflection point more precise. This is a smooth approximation of the heaviside step function. https://en.wikipedia.org/wiki/Logistic_function

Theseus.maxPenalty — Method

Penalizes values in vector that are above a threshold

Theseus.minPenalty — Method

Penalizes values in vector that are below a threshold

Theseus.pairDist — Method

Find the distance between all pairs of points in a point set. Returns a strictly lower triangular matrix.

Theseus.rigidSetCompare — Method

Compare the distance between all pairs of points in a target point set and the distance between all pairs of points in a form found point set.

Theseus.softmax — Method

Softmax is a generalized version of the logistic function. It returns a vector of probabilities that sum to 1.

Theseus.softplus — Method

Softplus is a smooth approximation of the ReLU function. A sharpness parameter k may be included to make the inflection point more precise.

x is the input, b is the inflection point bias, k is the sharpness parameter for the barrier slope.

negative k raises a barrier on the left side of the inflection point. positive k raises a barrier on the right side of the inflection point.

Theseus.target_xy — Method

Minimize the distance between the x and y coordinates of the target nodes and their corresponding nodes in the form found network. Equal to targeting a plan projection of the target nodes. Useful if the target geometry variation is dominated by the x and y coordinates.

Theseus.target_xyz — Method

Minimze distances between selected target nodes and their corresponding nodes in the form found network.